restructure the file path for ratingsReviews helpers
